requirements.txt
-f https://download.pytorch.org/whl/cu118/torch_stable.html
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.1

rdflib==6.3.2
SPARQLWrapper==2.0.0
elasticsearch[async]==8.6.1
networkx==3.0
EoN==1.1
matplotlib==3.7.1
pandas==1.5.3
nltk==3.8.1
tqdm==4.65.0
urllib3==1.26.15
requests==2.28.2
python-Levenshtein==0.20.9
Unidecode==1.3.6
tabulate==0.9.0
transformers==4.27.4
accelerate==0.18.0
evaluate==0.4.0
rouge_score==0.1.2
sacrebleu==2.3.1
scikit-learn==1.2.2
umap-learn[plot]==0.5.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
